# Optimize parsing performance