### Overview

This optimization avoids unnecessary object allocations by not taking snapshots of the parsing state when the size of the relevant nodes is `1`. Parsing time becomes about 48.6 to 57.1% of the original.

### Benchmark Results

#### Before Optimization f20bede62a516a11a468d27d3f1adde2085762bc

| Benchmark                            | Mode | Cnt |    Score |   Error | Units |
| :----------------------------------- | :--- | --: | -------: | ------: | :---- |
| `ParsingBenchmarks.parse_`           | avgt |  25 |  102.391 | ± 1.121 | ns/op |
| `ParsingBenchmarks.parse_a1i`        | avgt |  25 |  322.832 | ± 3.665 | ns/op |
| `ParsingBenchmarks.parse_c`          | avgt |  25 |  102.086 | ± 1.063 | ns/op |
| `ParsingBenchmarks.parse_k1i`        | avgt |  25 |  338.494 | ± 2.231 | ns/op |
| `ParsingBenchmarks.parse_l123456789` | avgt |  25 | 1014.364 | ± 9.041 | ns/op |

#### After Optimization f4d27ab04177908254b9b5d058a3ac6f9810a105

| Benchmark                            | Mode | Cnt |   Score |   Error | Units |
| :----------------------------------- | :--- | --: | ------: | ------: | :---- |
| `ParsingBenchmarks.parse_`           | avgt |  25 |  52.451 | ± 0.291 | ns/op |
| `ParsingBenchmarks.parse_a1i`        | avgt |  25 | 164.723 | ± 1.224 | ns/op |
| `ParsingBenchmarks.parse_c`          | avgt |  25 |  52.949 | ± 0.291 | ns/op |
| `ParsingBenchmarks.parse_k1i`        | avgt |  25 | 193.371 | ± 1.875 | ns/op |
| `ParsingBenchmarks.parse_l123456789` | avgt |  25 | 492.890 | ± 5.163 | ns/op |